Update patch set 21
Patch Set 21:

just wanted to mention that this needs to expose the equivalent ability of passing multiple environments with -e to heat stack-create. this is how satellite support is implemented if you look at:
https://review.gerrithub.io/#/c/231978/

It won't just be satellite integration that will work that way, but pretty much any time a user wants to use a non-default resource registry, so for cloudforms integration and the network isolation configuration, etc.

Patch-set: 21
Label: Code-Review=0