Use uuid4 instead of uuid1 as a base for passwords
uuid1 is intended to be uniq but not unpredictable. Knowing the
mac address of the host generating the uuid and what time it was
roughly generated at essentially reduces the parts of this uuid
that have to be guessed from 128 bits down to less then 48, still
a big number but there is no reason not to use the bigger of the
two.

Change-Id: Ib21c46dddbbd5d0e3aaaa019ba6d23aaaaa82e12