Update patch set 10
Patch Set 10: Code-Review-2

So, I think we need to refactor this a bit before we land it. The way we do parameters, I think we overwrite anything in Tuskar. Really, we should document what users need to send to Tuskar and then deploy what is in Tuskar.

We could perhaps add some of these defaults if they are not set in Tuskar yet.

Patch-set: 10