Update patch set 1
Patch Set 1: Code-Review-1

Based on the explanation in the review Giulio linked, I think I would prefer to go with --cinder-lvm.  Otherwise we may have issues if/when we have other drivers that also use iscsi (if there are any, I will admit I don't know).

I realize this creates a mismatch between the param name and the thing we're setting in the Heat template, but I think we can live with that for now and fix the template name later if we want.  Once this param goes into the CLI it becomes part of our API contract and we're stuck with it.

/2 cents

Patch-set: 1
Label: Code-Review=-1