Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Typo inline.

Also, this seems problematic to me because it's run unconditionally, AFAICT.  In the old script it was only run on a stack-create, but now if you use this to stack-update it will try to re-run all the initialization again, which will either fail or make a mess of your overcloud.

What was the problem with having the post-config as a separate step?  I actually kind of liked that, TBH.

Patch-set: 1
Reviewer: Ben Nemec <1003493@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1