Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I played around with this a bit, and it seems that we will not be able to use this parameter approach. It looks like parameters are only passed to the top level template, and not to any extra templates pulled in later. It would be helpful to have someone more familiar with Heat/TripleO confirm this.

If this turns out to be the case, I see two alternatives:

1. Use the CLI args to write an environment file and pass that in to Heat.
2. Document how/where to create the environment file by the user.

I am leaning towards option 2, because it reduces CLI options (we would just need rhel_reg boolean), and it is also more flexible (I only elected to put CLI options for the required parameters, but there are quite a few other ones).

Patch-set: 3
Label: Code-Review=-1