Align -e option with heatclient
Currently the long option and help text for the -e option is inconsistent
wrt heatclient - this option allows passing additional heat environment
files, but the current long option and help use the ambiguous term
extra templates, which could be confused for HOT templates.

The new long form is --environment-file, the -e short option is unchanged.

Change-Id: I7eff8965289076f93239c598fd21b537818bfe6e