Update patch set 5
Patch Set 5: Code-Review+1

Looks good.

> Just a note - default timeout is not obvious from this patch.

That's a hard thing to expose via the CLI, because it depends on the heat.conf configuration, which is not currently exposed via the API.

It defaults to 1 hour, but I guess we may want to increase that to match the undercloud token expiry which is 4 hours.

Patch-set: 5
Reviewer: Steven Hardy <1004023@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1