Use KeystoneAdminVip to obtain the keystone_ip
This patch updates overcloud deploy so that it obtains
the keystone admin IP address from a new KeystoneAdminVip
output parameter from Heat. This parameter is meant to
de-couple the Keystone Admin API VIP and should allow
us to make it easier to bootstrap keystone endpoints
when using network isolation on a minimal undercloud
setup (no extra networks are present).

Change-Id: I9693308412f275474ca0fbed07366c03fdc65126
Depends-On: I2bd3e61570151e2faeee14ee09b03ad0b3208cc1 (t-h-t)