Don't set storage params when not deploying Ceph
This is an attempt to partially fix a situation where oscplugin always
sets some stack parameters and they cannot be overriden [1].

This fix only affects the situation when Ceph is not being deployed (or
at least when Ceph is not managed by RDO Manager).

When Ceph *is* being deployed by RDO Manager, it's still assumed that it
will be the backend for both Cinder and Glance.

A proper fix needs to be implemented later (e.g. using an environment
file instead of direct parameters to pass parameters from oscplugin to
heat).

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1247585

Change-Id: I1479f1b74785adba03cf0b982c46e6763f2bc4d2