Don't assume storage backend choices even when using Ceph
This prevents the CLI from hard-wiring storage backend choices for
Cinder and Glance even when OSP-d is deploying Ceph.

This means we now allow user to select combinations like "Glance on
Swift, Cinder on Ceph" but it also means that user has to pass an
environment file with these settings *always*, even for the typical
"Ceph for everything" scenario. So we gain flexibility but lose "smart
behavior".

Change-Id: Ie0f6c763ea9cdeff60b6464935bd387daa285992