Add support for building ironic-python-agent ramdisk
Most upstream development in Ironic is happening on the agent
drivers rather than the pxe drivers. As such, we need to start
moving towards using the agent based drivers.

This patch is the first step in that direction. It allows building
the DIB based agent ramdisk from the openstack image build
command. It currently will not build in any of the defaults, but
this will allow us to set up CI that is building and using the
agent ramdisk in place of the pxe one.

Change-Id: Idc4486ad55fecf6de16b2855d2d80cbb97f1ab2a