WIP: Use parameter_defaults instead of parameters
This patch updates the overcloud deployment so that it
writes out a custom Heat environment to use parameter_defaults
instead of parameters for the stack creation. The parameter
defaults environment is placed first in the environments files
array so that end users can easily override any values with
settings in external Heat environments.

The motiviation here is to allow oscplugin to work in a wider
variety of environments instead of just forcing all users
to use the hard coded RDO defaults for things. Ultimately
it would be nice if the RDO defaults would actually live
outside of this project thus making the provisioning tool
more generically within the upstream TripleO community.
This is a step in that direction.

Functionally this works with --templates. WIP because
this still needs unit test updates.

Change-Id: I2722e118abde1cea6210148b16ac1ed3d9f235d8