Update patch set 6
Patch Set 6:

(1 comment)

Replied to the concern about where it should live on the mailing list.

As for the name, the problem is that it's very different from configuring tempest. It's similar to eg. write_overcloudrc except it might need to rerun on scaling if more features (eg. multiple cinder backends) will be added to rdomanager. I am planning to replace the documentation in instack-undercloud, I hope that will clarify things.

Patch-set: 6
Label: Code-Review=0