`openstack undercloud validate capabilities` plugin
The `undercloud validate capabilities` command takes the same scale and
flavor arguments given to the `overcloud deploy` command such as
`--control-scale`, and verifies the chain of profiles and boot options
that are necessary for a successful deploy. Where possible, it also
provides the CLI command to remedy the detected problem.

Ideally, this will save users from kicking off a deploy and diagnosing a
stack failure due to nova having insufficient nodes to schedule
instances.