Update patch set 5
Patch Set 5: Code-Review-1

The ironic-agent element does the ramdisk creation, so we should not be using the ramdisk element. This is also a liberty+ target, so we can probably wait to merge in to the openstack namespace.

Patch-set: 5
Reviewer: John Trowbridge <1003529@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1