Update patch set 3
Patch Set 3:

The reason I think we need to change the docs is that this breaks virt installs.  We really should have documented this before for baremetal deployments, but at least they still worked if you forgot to set it.  Virt will just blow up now if you try to launch an instance.

Patch-set: 3