Update patch set 1
Patch Set 1: Code-Review-1

(8 comments)

A few things:
-Please don't use sys.exit.  That will be really bad behavior when this gets moved into a library (or worse, behind a REST API).
-Would prefer not to use print for the same reason.  self.log should do what we need.
-Unit tests would be helpful.  This is non-trivial logic and we can't reasonably visually inspect the code for correctness.

Also, just a thought, but can we run this as part of the deployment instead of a separate step?  That would cut down on some of the redundancy in the options and then we wouldn't have to try to keep this in sync as params change (which they will in Liberty).

Alternatively, I'm wondering if this should be an overcloud command.  Given that it takes many of the same parameters as the deploy, it would be nice if users could just s/validate/deploy/ to get the correct deployment command after validation.  That would require even more duplication of params, but it would also potentially allow more validation in the future.  This actually makes me lean even more heavily toward just making it a deploy step though.

Patch-set: 1
Reviewer: Ben Nemec <1003493@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1