Update patch set 1
Patch Set 1:

(3 comments)

Thanks for the feedback. 

The reasoning behind making it an undercloud command is because it acts on existing configuration in the undercloud nova/ironic/etc. If you think it's better as an overcloud command, then I'll do that instead. 

As for the reasoning for having it as a separate step, I just thought it'd be a lower risk change than adding these checks to the deploy command. 

If I do add them to the deploy command, would you want the errors (e.g. flavor does not exist) to fail the command and warnings to just log? I think a good way to do it would be to exit (not using sys.exit) for errors and maybe throw up a confirmation dialog for warnings. Something like "there are N warnings about this configuration, logged above. Would you like to continue anyways" and have a --ignore-warnings option to continue regardless.

Patch-set: 1
Label: Code-Review=0