Update patch set 2
Patch Set 2:

@Brad: re: "Tests?"
This is a good point. Sure - I will add some tests for this (if this was your question). Also I wonder what is a test policy on this project? After brief check of recent patches it seems that about 50% of them miss tests.

Is the overall consensus that anything not having tests is automatically -1?

Patch-set: 2
Reviewer: Jan Provaznik <1003855@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0