Update patch set 3
Patch Set 3:

Sigh, gerrithub ate my original version of this comment.

Basically, yes, booting an instance on the overcloud won't work in a virt setup unless you pass --libvirt-type qemu (or have nested virt, which we don't recommend).

Since it's going to be a lot of people's first exposure to the tool, that kind of gotcha is pretty unpleasant (which is why the default was qemu in the first place, although changes since then have made it a problem).

Basically I want a Virtual admonition in the overcloud deployment section of the docs that specifically calls out --libvirt-type qemu.

Patch-set: 3
Reviewer: Ben Nemec <1003493@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0